Test supported repo versions with gha

+++ spec/acceptance/replset_spec.rb
@@ -2,14 +2,19 @@
 
 require 'spec_helper_acceptance'
 
-if hosts.length > 1
+repo_version = ENV.fetch('BEAKER_FACTER_mongodb_repo_version', nil)
+repo_ver_param = "repo_version => '#{repo_version}'" if repo_version
+
+if hosts.length > 1 && supported_version?(default[:platform], repo_version)
   describe 'mongodb_replset resource' do
     after :all do
       # Have to drop the DB to disable replsets for further testing
       on hosts, %{mongo local --verbose --eval 'db.dropDatabase()'}
 
       pp = <<-EOS
-        class { 'mongodb::globals': }
+        class { 'mongodb::globals':
+          #{repo_ver_param}
+        }
         -> class { 'mongodb::server':
           ensure         => absent,
           package_ensure => absent,
@@ -27,7 +32,9 @@ if hosts.length > 1
 
     it 'configures mongo on both nodes' do
       pp = <<-EOS
-        class { 'mongodb::globals': }
+        class { 'mongodb::globals':
+          #{repo_ver_param}
+        }
         -> class { 'mongodb::server':
           bind_ip => '0.0.0.0',
           replset => 'test',
@@ -79,7 +86,9 @@ if hosts.length > 1
       on hosts, %{mongo local --verbose --eval 'db.dropDatabase()'}
 
       pp = <<-EOS
-        class { 'mongodb::globals': }
+        class { 'mongodb::globals':
+          #{repo_ver_param}
+        }
         -> class { 'mongodb::server':
           ensure => absent,
           package_ensure => absent,
@@ -97,18 +106,17 @@ if hosts.length > 1
 
     it 'configures mongo on both nodes' do
       pp = <<~EOS
-                class { 'mongodb::globals':
-                  version             => '2.6.9-1',
-                  manage_package_repo => true
-                } ->
-                class { 'mongodb::server':
-                  admin_username => 'admin',
-                  admin_password => 'password',
-                  auth           => true,
-                  bind_ip        => '0.0.0.0',
-                  replset        => 'test',
-                  keyfile        => '/var/lib/mongodb/mongodb-keyfile',
-                  key            => '+dxlTrury7xtD0FRqFf3YWGnKqWAtlyauuemxuYuyz9POPUuX1Uj3chGU8MFMHa7
+        class { 'mongodb::globals':
+          #{repo_ver_param}
+        } ->
+        class { 'mongodb::server':
+          admin_username => 'admin',
+          admin_password => 'password',
+          auth           => true,
+          bind_ip        => '0.0.0.0',
+          replset        => 'test',
+          keyfile        => '/var/lib/mongodb/mongodb-keyfile',
+          key            => '+dxlTrury7xtD0FRqFf3YWGnKqWAtlyauuemxuYuyz9POPUuX1Uj3chGU8MFMHa7
         UxASqex7NLMALQXHL+Th4T3dyb6kMZD7KiMcJObO4M+JLiX9drcTiifsDEgGMi7G
         vYn3pWSm5TTDrHJw7RNWfMHw3sHk0muGQcO+0dWv3sDJ6SiU8yOKRtYcTEA15GbP
         ReDZuHFy1T1qhk5NIt6pTtPGsZKSm2wAWIOa2f2IXvpeQHhjxP8aDFb3fQaCAqOD
@@ -124,11 +132,10 @@ if hosts.length > 1
         YG/QX0BmltAni8owgymFuyJgvr/gaRX4WHbKFD+9nKpqJ3ocuVNuCDsxDqLsJEME
         nc1ohyB0lNt8lHf1U00mtgDSV3fwo5LkwhRi6d+bDBTL/C6MZETMLdyCqDlTdUWG
         YXIsJ0gYcu9XG3mx10LbdPJvxSMg'
-
-                }
-                if $::osfamily == 'RedHat' {
-                  include mongodb::client
-                }
+        }
+        if $::osfamily == 'RedHat' {
+          include mongodb::client
+        }
       EOS
 
       apply_manifest_on(hosts.reverse, pp, catch_failures: true)
@@ -137,43 +144,11 @@ if hosts.length > 1
 
     it 'sets up the replset with puppet' do
       pp = <<~EOS
-                class { 'mongodb::globals':
-                  version             => '2.6.9-1',
-                  manage_package_repo => true
-                } ->
-                class { 'mongodb::server':
-                  create_admin   => true,
-                  admin_username => 'admin',
-                  admin_password => 'password',
-                  auth           => true,
-                  bind_ip        => '0.0.0.0',
-                  replset        => 'test',
-                  keyfile        => '/var/lib/mongodb/mongodb-keyfile',
-                  key            => '+dxlTrury7xtD0FRqFf3YWGnKqWAtlyauuemxuYuyz9POPUuX1Uj3chGU8MFMHa7
-        UxASqex7NLMALQXHL+Th4T3dyb6kMZD7KiMcJObO4M+JLiX9drcTiifsDEgGMi7G
-        vYn3pWSm5TTDrHJw7RNWfMHw3sHk0muGQcO+0dWv3sDJ6SiU8yOKRtYcTEA15GbP
-        ReDZuHFy1T1qhk5NIt6pTtPGsZKSm2wAWIOa2f2IXvpeQHhjxP8aDFb3fQaCAqOD
-        R7hrimqq0Nickfe8RLA89iPXyadr/YeNBB7w7rySatQBzwIbBUVGNNA5cxCkwyx9
-        E5of3xi7GL9xNxhQ8l0JEpofd4H0y0TOfFDIEjc7cOnYlKAHzBgog4OcFSILgUaF
-        kHuTMtv0pj+MMkW2HkeXETNII9XE1+JiZgHY08G7yFEJy87ttUoeKmpbI6spFz5U
-        4K0amj+N6SOwXaS8uwp6kCqay0ERJLnw+7dKNKZIZdGCrrBxcZ7wmR/wLYrxvHhZ
-        QpeXTxgD5ebwCR0cf3Xnb5ql5G/HHKZDq8LTFHwELNh23URGPY7K7uK+IF6jSEhq
-        V2H3HnWV9teuuJ5he9BB/pLnyfjft6KUUqE9HbaGlX0f3YBk/0T3S2ESN4jnfRTQ
-        ysAKvQ6NasXkzqXktu8X4fS5QNqrFyqKBZSWxttfJBKXnT0TxamCKLRx4AgQglYo
-        3KRoyfxXx6G+AjP1frDJxFAFEIgEFqRk/FFuT/y9LpU+3cXYX1Gt6wEatgmnBM3K
-        g+Bybk5qHv1b7M8Tv9/I/BRXcpLHeIkMICMY8sVPGmP8xzL1L3i0cws8p5h0zPBa
-        YG/QX0BmltAni8owgymFuyJgvr/gaRX4WHbKFD+9nKpqJ3ocuVNuCDsxDqLsJEME
-        nc1ohyB0lNt8lHf1U00mtgDSV3fwo5LkwhRi6d+bDBTL/C6MZETMLdyCqDlTdUWG
-        YXIsJ0gYcu9XG3mx10LbdPJvxSMg'
-                }
-                if $::osfamily == 'RedHat' {
-                  include mongodb::client
-                }
-                mongodb_replset { 'test':
-                  auth_enabled => true,
-                  members      => [#{hosts.map { |x| "'#{x}:27017'" }.join(',')}],
-                  before       => Mongodb_user['admin']
-                }
+        mongodb_replset { 'test':
+          auth_enabled => true,
+          members      => [#{hosts.map { |x| "'#{x}:27017'" }.join(',')}],
+          before       => Mongodb_user['admin']
+        }
       EOS
       apply_manifest_on(hosts_as('master'), pp, catch_failures: true)
       apply_manifest_on(hosts_as('master'), pp, catch_changes: true)
